setup desktop Tap Handler, renamed mobile events to include word share. setup all current mobile functions.  Accel and Gyro not fully functional atm.
