+++ client/tapmobile.ejs
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="mobile.css">
   </head>
   <body>
-    <button type="button" name="button" onclick="frontEndEcho.mobileTap()">Tap Here</button>
+    <button type="button" name="button" onclick="frontEndEcho.mobileTapShare()">Tap Here</button>
     <div></div>
     <h1></h1>
     <div></div>
